[REFACTOR] totalCount 타입을 int에서 long으로 변경

+++ src/main/java/sopt/makers/authentication/adapter/in/web/dto/user/response/UserResponse.java
@@ -62,12 +62,11 @@ public final class UserResponse {
   }
 
   public record PaginatedUserProfiles(
-      List<UserProfileAndActivity> profiles, boolean hasNext, int totalCount) {
+      List<UserProfileAndActivity> profiles, boolean hasNext, long totalCount) {
     public static PaginatedUserProfiles from(
         GetUserProfileUsecase.PaginatedUserProfiles paginatedData) {
       List<UserProfileAndActivity> profiles =
           paginatedData.profiles().stream().map(UserProfileAndActivity::from).toList();
-
       return new PaginatedUserProfiles(
           profiles, paginatedData.hasNext(), paginatedData.totalCount());
     }
